requirements.txt
tensorflow==2.19.0
opencv-python
numpy
pillow
matplotlib
fastapi
uvicorn[standard]
python-multipart 